# queries/admins_queries.sql.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1
// source: admins_queries.sql

package queries

import (
	"context"
)

const countAdmins = `-- name: CountAdmins :one
SELECT COUNT(*) FROM admins
`

func (q *Queries) CountAdmins(ctx context.Context) (int64, error) {
	row := q.db.QueryRowContext(ctx, countAdmins)
	var count int64
	err := row.Scan(&count)
	return count, err
}

const createAdmin = `-- name: CreateAdmin :one
INSERT INTO admins (email, password_hash)
VALUES (?, ?)
RETURNING id, email, created_at
`

type CreateAdminParams struct {
	Email        string `json:"email"`
	PasswordHash string `json:"password_hash"`
}

type CreateAdminRow struct {
	ID        int64  `json:"id"`
	Email     string `json:"email"`
	CreatedAt string `json:"created_at"`
}

func (q *Queries) CreateAdmin(ctx context.Context, arg CreateAdminParams) (CreateAdminRow, error) {
	row := q.db.QueryRowContext(ctx, createAdmin, arg.Email, arg.PasswordHash)
	var i CreateAdminRow
	err := row.Scan(&i.ID, &i.Email, &i.CreatedAt)
	return i, err
}

const createAdminSession = `-- name: CreateAdminSession :exec
INSERT INTO admin_sessions (token_hash, admin_id, expires_at)
VALUES (?, ?, ?)
`

type CreateAdminSessionParams struct {
	TokenHash []byte `json:"token_hash"`
	AdminID   int64  `json:"admin_id"`
	ExpiresAt string `json:"expires_at"`
}

func (q *Queries) CreateAdminSession(ctx context.Context, arg CreateAdminSessionParams) error {
	_, err := q.db.ExecContext(ctx, createAdminSession, arg.TokenHash, arg.AdminID, arg.ExpiresAt)
	return err
}

const getAdminBySession = `-- name: GetAdminBySession :one
SELECT admins.id, admins.email, admins.created_at
FROM admin_sessions
JOIN admins ON admins.id = admin_sessions.admin_id
WHERE admin_sessions.token_hash = ?
  AND admin_sessions.expires_at > ?
`

type GetAdminBySessionParams struct {
	TokenHash []byte `json:"token_hash"`
	ExpiresAt string `json:"expires_at"`
}

type GetAdminBySessionRow struct {
	ID        int64  `json:"id"`
	Email     string `json:"email"`
	CreatedAt string `json:"created_at"`
}

func (q *Queries) GetAdminBySession(ctx context.Context, arg GetAdminBySessionParams) (GetAdminBySessionRow, error) {
	row := q.db.QueryRowContext(ctx, getAdminBySession, arg.TokenHash, arg.ExpiresAt)
	var i GetAdminBySessionRow
	err := row.Scan(&i.ID, &i.Email, &i.CreatedAt)
	return i, err
}

const getAdminCredentialsByEmail = `-- name: GetAdminCredentialsByEmail :one
SELECT id, email, password_hash, created_at
FROM admins
WHERE email = ? COLLATE NOCASE
`

func (q *Queries) GetAdminCredentialsByEmail(ctx context.Context, email string) (Admin, error) {
	row := q.db.QueryRowContext(ctx, getAdminCredentialsByEmail, email)
	var i Admin
	err := row.Scan(
		&i.ID,
		&i.Email,
		&i.PasswordHash,
		&i.CreatedAt,
	)
	return i, err
}